LayoutTests/fast/animation/request-animation-frame-throttling-outside-viewport-expected.txt

Test that requestAnimationFrame gets the right throttling in an iframe when when it's outside the viewport.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS framesPerSecond > 0 is true
PASS iframeFramesPerSecond == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
